requirements.txt
Django==4.2.7
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
Pillow==10.1.0
django-allauth==0.59.0
django-ckeditor==6.7.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
